requirements.txt
Flask~=3.0.2
numpy~=1.26.4
tensorflow~=2.15.0
Pillow==10.0.1
requests==2.31.0
gunicorn